refactor: 정적 리소스를 처리하는 Contollre의 이름을 DefaultController로 변경
