+++ tomcat/src/main/java/com/techcourse/controller/mapper/RequestMapper.java
@@ -1,14 +1,14 @@
-package org.apache.catalina;
+package com.techcourse.controller.mapper;
 
 import com.techcourse.controller.LoginController;
 import com.techcourse.controller.NotFoundController;
 import com.techcourse.controller.RegisterController;
-import com.techcourse.controller.ResourceController;
 import com.techcourse.controller.RootController;
 import com.techcourse.service.RegisterService;
 import com.techcourse.service.UserService;
 import java.util.HashMap;
 import java.util.Map;
+import org.apache.catalina.controller.DefaultController;
 import org.apache.coyote.Controller;
 import org.apache.coyote.http11.request.HttpRequest;
 
@@ -27,7 +27,7 @@ public class RequestMapper {
 
     public static Controller getController(HttpRequest request) {
         if (request.isStaticResource()) {
-            return new ResourceController();
+            return new DefaultController();
         }
 
         return handlerMap.getOrDefault(request.getUri(), new NotFoundController());
